change demicalfield to floatfield

+++ soccersite/map/models.py
@@ -26,6 +26,6 @@ class HighSchool(models.Model):
     instiution      = models.CharField(max_length = 100)
     stateOrProvince = models.CharField(max_length = 20)
     country         = models.CharField(max_length = 50)
-    latitude        = models.DecimalField(max_digits = 10)
-    longitude       = models.DecimalField(max_digits = 10)
+    latitude        = models.FloatField()
+    longitude       = models.FloatField()
     schoolType      = models.CharField(max_length = 20)
